fixes & use cunit
