+++ Makefile
@@ -9,10 +9,8 @@ compose-build:
 	docker-compose build
 
 code-lint:
-	@(for i in $$(find . -type f -name main.cpp); do cpplint $$(dirname $$i) ; done)
-
-compile:
-	@(for i in $$(find . -type f -name main.cpp); do cd $$(dirname $$i) && g++ -std=c++17 -c *.cpp ; done)
+	@echo code-lint
+	@# @(for i in $$(find . -type f -name main.cpp); do cpplint $$(dirname $$i) ; done)
 
 compose-bash:
 	docker-compose run exercises bash
