+++ modules/10-basics/10-hello-world/main.c
@@ -5,5 +5,5 @@ int main(void)
     // BEGIN
     printf("Hello, World!");
     // END
-    return 1;
+    return 0;
 }
