+++ modules/10-basics/10-hello-world/test.sh
@@ -2,6 +2,6 @@ ACTUAL=`tcc -run main.c`
 EXPECTED="Hello, World!"
 
 if [ "$ACTUAL" != "$EXPECTED" ]; then
-    echo "Expected '$EXPECTED', but actual was '$ACTUAL'"
+    echo "Expected:\n$EXPECTED\nReceived:\n$ACTUAL"
     exit 1
 fi
