+++ modules/10-basics/20-variables/test.sh
@@ -17,6 +17,6 @@ EXPECTED='0 : 273
 300 : 573';
 
 if [ "$ACTUAL" != "$EXPECTED" ]; then
-    echo "Expected '$EXPECTED', but actual was '$ACTUAL'"
+    echo "Expected:\n$EXPECTED\nReceived:\n$ACTUAL"
     exit 1
 fi
