+++ modules/10-basics/30-floats/test.sh
@@ -19,6 +19,6 @@ EXPECTED='0.00 : 0.00
 300.00 : 186.30';
 
 if [ "$ACTUAL" != "$EXPECTED" ]; then
-    echo "Expected '$EXPECTED', but actual was '$ACTUAL'"
+    echo "Expected:\n$EXPECTED\nReceived:\n$ACTUAL"
     exit 1
 fi
