+++ modules/10-basics/40-for/test.sh
@@ -19,7 +19,6 @@ EXPECTED='0.00 : -17.78
 300.00 : 148.89';
 
 if [ "$ACTUAL" != "$EXPECTED" ]; then
-    echo "Expected '$EXPECTED', but actual was '$ACTUAL'"
+    echo "Expected:\n$EXPECTED\nReceived:\n$ACTUAL"
     exit 1
 fi
-
