--- modules/10-basics/40-functions/cases.txt
@@ -1,3 +1,0 @@
-3200 2927
--273 -546
-0 -273
