+++ modules/10-basics/50-functions/Makefile
@@ -1,0 +1,2 @@
+test:
+	@tcc -lcunit -run test.c
