+++ modules/10-basics/50-functions/main.c
@@ -1,0 +1,8 @@
+// BEGIN
+int kelvin_to_celcius(int kelvin) {
+
+    int celcius;
+    celcius = kelvin - 273;
+    return celcius;
+}
+// END
