--- modules/10-basics/50-if-else/Makefile
@@ -1,2 +1,0 @@
-test:
-	@ ./test.sh
