--- modules/10-basics/50-if-else/description.ru.yml
@@ -1,13 +1,0 @@
----
-name: Условие if/else
-theory: |
-  Научимся использовать условия и выполним классическое упражнение Fizz-Buzz.
-
-  <div class="ratio ratio-16x9">
-    <iframe src="//player.vimeo.com/video/128854221?api=1&player_id=video1" id="video1"></iframe>
-  </div>
-
-instructions: |
-  Напишите функцию `kelvin_to_celcius()`, которая принимает температуру в Кельвинах и возвращает соответствующее значение в Цельсиях.
-
-  Функция должна принимать и возвращать `int`.
