+++ modules/10-basics/60-if-else/Makefile
@@ -1,0 +1,2 @@
+test:
+	@./test.sh
