+++ modules/10-basics/60-if-else/description.ru.yml
@@ -1,0 +1,25 @@
+---
+name: Условие if/else
+theory: |
+  Научимся использовать условия и выполним классическое упражнение Fizz-Buzz.
+
+  <div class="ratio ratio-16x9">
+    <iframe src="//player.vimeo.com/video/128854221?api=1&player_id=video1" id="video1"></iframe>
+  </div>
+
+instructions: |
+  Напишите функцию `fizzbuzz()`, которая принимает на вход в качестве параметра число `limit` типа `int`. В функции должен быть цикл, который проверяет числа от 1 до `limit`, и:
+
+  * если число делится на 3 без остатка, то на экран выводится слово 'Fizz'
+  * если число делится на 5 без остатка, то на экран выводится слово 'Buzz'
+  * если число делится и на 3, и на 5 без остатка, то на экран выводится слово 'FizzBuzz'
+  * в противном случае на экран выводится само число
+
+  Функция должна печатать результат на экран.
+
+  ```
+  fizzbuzz(5);
+  // => '1 2 Fizz 4 Buzz'
+  fizzbuzz(8);
+  // => '1 2 Fizz 4 Buzz Fizz 7 8'
+  ```
