+++ modules/10-basics/60-if-else/main.c
@@ -1,24 +1,16 @@
 #include <stdio.h>
 
-void fizzbuzz(int limit);
-
-int main(void)
-{
-    fizzbuzz(20);
-    return 0;
-}
-
 // BEGIN
 void fizzbuzz(int limit)
 {
     int i;
 
     for (i=1; i<=limit; ++i) {
-        if (!(i % 15))
+        if (i % 15 == 0)
             printf("FizzBuzz");
-        else if (!(i % 3))
+        else if (i % 3 == 0)
             printf("Fizz");
-        else if (!(i % 5))
+        else if (i % 5 == 0)
             printf("Buzz");
         else
             printf("%d", i);
