+++ modules/10-basics/60-if-else/test.c
@@ -1,0 +1,7 @@
+#include "main.c"
+
+int main(void)
+{
+    fizzbuzz(20);
+    return 0;
+}
