+++ modules/10-basics/60-if-else/test.sh
@@ -1,7 +1,7 @@
-ACTUAL=`tcc -run main.c`
+ACTUAL=`tcc -run test.c`
 EXPECTED='1 2 Fizz 4 Buzz Fizz 7 8 Fizz Buzz 11 Fizz 13 14 FizzBuzz 16 17 Fizz 19 Buzz';
 
 if [ "$ACTUAL" != "$EXPECTED" ]; then
-    echo "Expected '$EXPECTED', but actual was '$ACTUAL'"
+    echo "Expected:\n$EXPECTED\nReceived:\n$ACTUAL"
     exit 1
 fi
