--- modules/10-basics/60-switch/Makefile
@@ -1,2 +1,0 @@
-test:
-	@ ./test.sh
