+++ modules/10-basics/70-switch/main.c
@@ -1,19 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void number_printer(int number);
-
-int main(int argc, char **argv)
-{
-  int num = atoi(argv[1]);
-  number_printer(num);
-  return 0;
-}
-
 // BEGIN
 void number_printer(int number)
 {
-
   switch(number)
   {
     case 0:
