+++ modules/10-basics/70-switch/test.c
@@ -1,0 +1,8 @@
+#include "main.c"
+
+int main(int argc, char **argv)
+{
+  int num = atoi(argv[1]);
+  number_printer(num);
+  return 0;
+}
