+++ modules/10-basics/70-switch/test.sh
@@ -3,11 +3,11 @@
 while read case; do
     set -- $case
 
-    ACTUAL=`tcc -run main.c $1`
+    ACTUAL=`tcc -run test.c $1`
     EXPECTED=$2
 
     if [ "$ACTUAL" != "$EXPECTED" ]; then
-        echo "Expected '$EXPECTED', but actual was '$ACTUAL'"
+        echo "Expected:\n$EXPECTED\nReceived:\n$ACTUAL"
         exit 1
     fi
